packages/backend/syntek-auth/pyproject.toml
[project]
name = "syntek-auth"
version = "0.4.0"
description = "Authentication module for Syntek — MFA, passkeys, OAuth, Argon2id"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0",
    "strawberry-graphql[django]>=0.307",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.9",
    "factory-boy>=3.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/syntek_auth"]
